+++ fol/constitutive_material_models/base.py
@@ -1,3 +1,8 @@
+"""
+ Authors: Rishabh Arora, https://github.com/rishabharora236-cell
+ Date: Dec, 2025
+ License: FOL/LICENSE
+"""
 from abc import ABC, abstractmethod
 from typing import Optional, Any, Tuple
 import jax.numpy as jnp
@@ -19,24 +24,4 @@ class BaseConstitutiveModel(ABC):
         - Small-strain plastic: (stress, new_state)
         - Large-strain plastic: (stress, new_state)
         """
-        pass
-    
-
-
-
-class PlasticityModel(BaseConstitutiveModel):
-    """Base for plasticity models (with history)"""
-    
-    @abstractmethod
-    def evaluate(self, strain: Array, state: Array) -> Tuple[Array, Array, Array]:
-        """
-        Returns:
-            stress: Cauchy stress
-            new_state: Updated state vector
-        """
-        pass
-    
-    @abstractmethod
-    def initial_state(self, dim: int = 3) -> Array:
-        """Must return initial state vector"""
         pass
